Feat: 10만 단위로 부하 테스트ì 구현중

+++ src/main/java/com/project/final_project/board/dto/BoardDTO.java
@@ -13,9 +13,9 @@ public class BoardDTO {
   String title;
   String content;
   Integer likeCount;
-  Integer commentCount;
+  Long commentCount;
 
-  public BoardDTO(Board board, Integer commentCount) {
+  public BoardDTO(Board board, Long commentCount) {
     this.boardId = board.getId();
     this.title = board.getTitle();
     this.content = board.getContent();
